community insert, list

+++ git/Springboot/pet/src/main/java/com/wproject/pet/repository/CommunityRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.wproject.pet.entity.Community;
 
-public interface CommunityRepository extends JpaRepository<Community, Integer>{
+public interface CommunityRepository extends JpaRepository<Community, Long>{
 	public Page<Community> findByBTitleContaining(String word, Pageable pageable);
 	public Page<Community> findByBContentContaining(String word, Pageable pageable);
